Use project dependencies, not maven dependencies
Prior to this commit it was tricky to get the gradle maven plugin to correctly
name each dependency in the various generated POMs. Because of this, each dependency
was specified as a maven dependency, rather than a project dependency, to tell the maven
plugin what everything should be called. This update changes a few things:

- Import and name projects in settings.gradle
- In addition to name, set moduleId and version at the project level
- Dynamically add kit project dependencies to the all-kits artifact
- Remove the need for gradle.properties in every kit project

Given the above, the project can always be built and uploaded with a single command,
even as the version number changes.